@@ -40,6 +40,7 @@
     "fs-extra": "4.0.0",
     "glob": "^7.1.2",
     "helmet": "3.6.1",
+    "intl": "^1.2.5",
     "inversify": "4.2.0",
     "inversify-binding-decorators": "3.0.0",
     "inversify-express-utils": "4.0.0",
@@ -58,12 +59,13 @@
     "require-from-string": "^1.2.1",
     "rxjs": "^5.4.3",
     "simple-encryptor": "^1.1.0",
+    "web-animations-js": "^2.3.1",
     "webpack": "3.4.1",
     "zone.js": "0.8.14"
   },
   "devDependencies": {
     "@angular/animations": "4.4.3",
-    "@angular/cli": "1.3.0-rc.3",
+    "@angular/cli": "^1.4.4",
     "@angular/common": "4.4.3",
     "@angular/compiler": "4.4.3",
     "@angular/compiler-cli": "4.4.3",
@@ -132,7 +134,7 @@
     "ts-loader": "2.3.1",
     "ts-node": "3.2.0",
     "tslint": "5.3.2",
-    "typescript": "2.4.0",
+    "typescript": "^2.5.3",
     "url-loader": "0.5.7",
     "webpack-dev-server": "2.5.1"
   }
