@@ -1,14 +1,18 @@
 import { Injectable } from '@angular/core'
 import { ActivatedRoute, Event, NavigationEnd, NavigationStart, Router } from '@angular/router'
-import { Observable, Subject } from 'rxjs/Rx'
+import { Observable } from 'rxjs/Observable'
+import { Subject } from 'rxjs/Subject'
+import 'rxjs/add/observable/of'
 import 'rxjs/add/operator/filter'
+import 'rxjs/add/operator/switchMap'
 import { UiService } from './ui.service'
 import { UserService } from './user.service'
 
 @Injectable()
 export class RouteStateService {
 
-    public previousUrl: string
+    public previousUrl = "/"
+    public routerEvents$: Observable<any>
 
     constructor(
         private router: Router,
@@ -18,12 +22,22 @@ export class RouteStateService {
     }
 
     public init() {
-        this.router.events
+        Observable.of(null)
+            .switchMap(x => this.router.events)
+            .filter(e => e instanceof NavigationStart)
             .subscribe(e => {
-                if (e instanceof NavigationEnd) {
-                    this.previousUrl = e.url
-                    console.log(e)
+                if (this.route.snapshot.firstChild
+                    && this.route.snapshot.firstChild.url
+                    && this.route.snapshot.firstChild.url.length) {
+                    this.previousUrl = this.route.snapshot.firstChild.url.map(segment => "/" + segment.path).join("")
+                } else {
+                    this.previousUrl = "/"
                 }
+                console.log("Previous URL", this.previousUrl)
             })
+    }
+
+    public refreshPage() {
+        window.location.reload()
     }
 }
