@@ -1,26 +1,26 @@
-import { Injectable, EventEmitter, ChangeDetectorRef } from '@angular/core'
+import { Injectable } from '@angular/core'
 import { Title } from '@angular/platform-browser'
-import { Observable, Subject } from 'rxjs/Rx'
-import { appConfig } from '@time/app-config'
-import { SimpleError } from '@time/common/http'
+import { Subject } from 'rxjs/Subject'
 
-export type FlashMessageType = 'success'|'error'|'info'|'warn'
-export interface IFlash {
-    type: string
-    message: string
-    timeout?: number
-}
+import { SimpleError, TimeHttpResponseInterceptor, TimeHttpService } from '@time/common/ng-modules/http'
+import { FlashMessageType, IFlash } from '../models/ui.models'
 
 @Injectable()
 export class UiService {
     public viewReady$ = new Subject<boolean>()
     public flash$ = new Subject<IFlash>()
+    public flashCancel$ = new Subject<any>()
 
     constructor(
         private titleService: Title,
-    ) {}
+        private timeHttpService: TimeHttpService,
+    ) {
+        this.timeHttpService.error$.subscribe(err => {
+            console.log("ERROR!!@#!@!@#!@#", err)
+        })
+    }
 
-    setTitle(title: string) {
+    public setTitle(title: string) {
         this.titleService.setTitle(title)
     }
 
@@ -30,19 +30,22 @@ export class UiService {
 	 * @param msg
 	 * @param {string} [t = success|error|info|warning]
 	 */
-    flash(message: string, type: FlashMessageType = 'info', timeout: number = 5000) {
+    public flash(message: string, type: FlashMessageType = 'info', timeout: number = 5000) {
         const data: IFlash = {
             type,
             message,
             timeout,
         }
         this.flash$.next(data)
+        setTimeout(() => {
+            this.flashCancel$.next()
+        }, timeout)
     }
 
 	/**
 	 * Display an error as a flash message
 	 */
-    flashError(error: SimpleError) {
+    public flashError(error: SimpleError) {
         this.flash(error.message, 'error')
     }
 }
