@@ -1,13 +1,16 @@
 import { HttpClient } from '@angular/common/http'
-import { EventEmitter, Inject, Injectable } from '@angular/core'
-import { CanActivate, CanActivateChild, Router } from '@angular/router'
-import { Observable, Subject } from 'rxjs/Rx'
+import { Injectable } from '@angular/core'
+import { Observable } from 'rxjs/Observable'
+import { Subject } from 'rxjs/Subject'
 
+import { CONSTANTS } from '@time/common/constants'
 import { IUser } from '@time/common/models/interfaces'
+import { TimeHttpService } from '@time/common/ng-modules/http'
 import { UtilService } from './util.service'
 
 @Injectable()
 export class UserService {
+    private jwtToken: string
     public user: IUser
     public user$: Subject<IUser> = new Subject<IUser>()
     public notifications = { sessions: 0, conversations: 0 }
@@ -15,61 +18,50 @@ export class UserService {
 
     constructor (
         private http: HttpClient,
-        private router: Router,
+        private timeHttpService: TimeHttpService,
         private util: UtilService,
-    ) {}
+    ) {
+        this.timeHttpService.auth$.subscribe(token => {
+            this.jwtToken = token
+            let base64: string = token.split('.')[1]
+            base64 = base64.replace('-', '+').replace('_', '/')
+            this.user = JSON.parse(window.atob(base64))
+            localStorage.setItem("authToken", this.jwtToken)
+        })
 
-    public signup(user): Observable<any> {
-        return this.http.post('/api/signup', user)
+        this.timeHttpService.error$.subscribe(err => {
+            if (err.status === CONSTANTS.HTTP.CLIENT_ERROR_forbidden || err.status === CONSTANTS.HTTP.CLIENT_ERROR_unauthorized) {
+                if (this.user) {
+                    this.doLogout()
+                    window.location.reload()
+                }
+            }
+        })
     }
 
-    public login(credentials: {email: string; password: string}): Observable<any> {
-        return this.http.post('/api/login/local', credentials)
+    public get token(): string {
+        return this.jwtToken || <string>this.util.getFromLocalStorage("authToken")
+    }
+
+    private doLogout(): void {
+
+    }
+
+    public signup(user): Observable<any> {
+        return this.http.post('/api/user/register', user)
+    }
+
+    public login(credentials: {email: string; password: string}): void {
+        // FOR TESTING
+        this.http.get('/api/user/login').subscribe(data => console.log(data))
+        // this.http.post('/api/user/login', credentials)
     }
 
     public logout(): Observable<any> {
-        return this.http.post('/api/logout', {})
+        return this.http.post('/api/user/logout', {})
     }
 
     public verifyEmail(token: string): Observable<IUser> {
         return this.http.get<IUser>(`/api/verify-email/${token}`)
     }
-
-    // edit(user: any, done?: (err?: any, user?: IUser) => void) {
-    //     this.http.post('/api/user/edit', user)
-    //         .map((res: Response) => res.json())
-    //         .catch(this.util.catchHttpError)
-    //         .subscribe(
-    //             user => {
-    //                 this.onLogin(user)
-    //                 if (done) done(null, user)
-    //             },
-    //             err => {
-    //                 if (err) done(err)
-    //                 this.util.handleError(err)
-    //             },
-    //         )
-    // }
-
-    // editPassword(newPassword: string, resetToken?: string, done?: (err?: any, user?: IUser) => void) {
-    //     this.http.post('/api/reset-password', {newPassword, resetToken})
-    //         .map((res: Response) => res.json())
-    //         .catch(this.util.catchHttpError)
-    //         .subscribe(
-    //             user => {
-    //                 this.onLogin(user)
-    //                 if (done) done(null, user)
-    //             },
-    //             err => {
-    //                 if (err) done(err)
-    //                 this.util.handleError("Password update failed")
-    //             },
-    //         )
-    // }
-
-    // forgotPassword(email: string): Observable<IUser> {
-    //     return this.http.post('/api/send-forgot-password', {email})
-    //         .map((res: Response) => res.json())
-    //         .catch(this.util.catchHttpError)
-    // }
 }
