@@ -1,12 +1,13 @@
-import { HttpClient, HttpErrorResponse, HttpParams } from '@angular/common/http'
+import { HttpClient, HttpErrorResponse, HttpHeaders, HttpParams } from '@angular/common/http'
 import { EventEmitter, Inject, Injectable } from '@angular/core'
-import { Observable, Subject } from 'rxjs/Rx'
+import { Observable } from 'rxjs/Observable'
+import { Subject } from 'rxjs/Subject'
 
 import { appConfig } from '@time/app-config'
 import { ProductSearch } from '@time/common/api-utils'
-import { SimpleError } from '@time/common/http'
 import { IProduct } from '@time/common/models/interfaces'
-import { UtilService } from '../../shared/services'
+import { SimpleError } from '@time/common/ng-modules/http'
+import { UserService, UtilService } from '../../shared/services'
 
 @Injectable()
 export class ProductService {
@@ -17,7 +18,7 @@ export class ProductService {
 
     constructor (
         private http: HttpClient,
-        private util: UtilService,
+        private userService: UserService,
     ) {
         this.products$ = this.productsSubject.asObservable()
         this.productsError$ = this.productsErrorSubject.asObservable()
@@ -25,17 +26,19 @@ export class ProductService {
 
     public get(page?: number, query?: ProductSearch.Body): void {
         const params = new HttpParams()
+        const headers = new HttpHeaders({authorization: "bearer " + this.userService.token})
+
         if (page) params.set("page", page + "")
         if (query) params.set("page", JSON.stringify(query))
 
-        setInterval(() => {
-            this.http.get("https://jsonplaceholder.typicode.com/posts")
-            // this.http.get(appConfig.client_url + "/api/products", { params })
-                .subscribe(
-                    (products: IProduct[]) => this.productsSubject.next(products),
-                    (error: SimpleError) => this.productsErrorSubject.next(error),
-                )
-        }, 2000)
+        console.log({ params, headers })
+
+        // this.http.get("https://jsonplaceholder.typicode.com/posts")
+        this.http.get(appConfig.client_url + "/api/products", { params, headers })
+            .subscribe(
+                (products: IProduct[]) => this.productsSubject.next(products),
+                (error: SimpleError) => this.productsErrorSubject.next(error),
+            )
     }
 
     public getOne(slug: string): void {
