@@ -1,14 +1,14 @@
-import { TimeClientPage } from './app.po';
+import { TimeClientPage } from './app.po'
 
 describe('time-client App', () => {
-  let page: TimeClientPage;
+  let page: TimeClientPage
 
   beforeEach(() => {
-    page = new TimeClientPage();
-  });
+    page = new TimeClientPage()
+  })
 
   it('should display welcome message', () => {
-    page.navigateTo();
-    expect(page.getParagraphText()).toEqual('Welcome to app!');
-  });
-});
+    page.navigateTo()
+    expect(page.getParagraphText()).toEqual('Welcome to app!')
+  })
+})
