@@ -2,6 +2,7 @@ import { HTTP_INTERCEPTORS } from '@angular/common/http'
 import { NgModule } from '@angular/core'
 import { BrowserModule } from '@angular/platform-browser'
 import { RouterModule } from '@angular/router'
+import { TimeHttpResponseInterceptor } from '@time/common/ng-modules/http/http-response.interceptor'
 import { SharedModule } from './shared/shared.module'
 
 import { ShopModule } from './shop/shop.module'
@@ -20,9 +21,7 @@ import { AppComponent } from './app.component'
         SiteModule,
         ShopModule,
     ],
-    providers: [
-
-    ],
+    providers: [],
     bootstrap: [AppComponent],
 })
 export class AppModule { }
