@@ -1,41 +1,27 @@
 import { Injectable } from '@angular/core'
-import { Title } from '@angular/platform-browser'
-import { Observable } from 'rxjs/Observable'
 import { Subject } from 'rxjs/Subject'
 import 'rxjs/add/observable/throw'
 import 'rxjs/add/operator/catch'
 import 'rxjs/add/operator/map'
+import { SeoService } from './seo.service'
 
 import { UiService } from './ui.service'
 
 @Injectable()
 export class UtilService {
 
-    private dateMidnightMillis: number
     public userError$ = new Subject<any>()
 
     constructor(
-        private titleService: Title,
         private ui: UiService,
+        private seo: SeoService,
     ) {}
 
     public setTitle(title: string): void {
-        this.titleService.setTitle(title + ' | A2EE')
+        this.seo.setTitle(title)
     }
 
-    public handleError(serverMsg: any, msg?: string): boolean {
-        if (typeof serverMsg === "string") {
-            if (serverMsg.charAt(0) === "{") serverMsg = JSON.parse(serverMsg).error
-        }
-        console.error(serverMsg)
-        if (this.ui && this.ui.flash) {
-            this.ui.flash(msg || serverMsg || "Oops! Something went wrong. Please try again.", "error")
-        }
-        if (serverMsg) return true
-        else return false
-    }
-
-    public getFromLocalStorage(key: string): any {
+    public getFromLocalStorage(key: string): object|string|undefined {
         const item = localStorage.getItem(key)
         if (item && item.length) {
             if (item.charAt(0) === '[' || item.charAt(0) === '{') {
@@ -45,7 +31,7 @@ export class UtilService {
                 return item
             }
         }
-        return null
+        return undefined
     }
 
     public gToOz(pK: number): number {
@@ -57,51 +43,5 @@ export class UtilService {
 
     public ozToG(pK: number): number {
         return pK * 28.3495
-    }
-
-    public getTimes(noAmPm?: boolean): Array<{name: string; value: number}> {
-        const options = []
-        const loop = (amPm: 'am'|'pm') => {
-            for (let i = 0; i < 48; i++) { // 15 minute increments
-                const fractionStr = (i / 4).toFixed(2).toString()
-                let timeStr: string
-                let timeStrBefore: string
-                let timeStrAfter: string
-                const halfDay: number = amPm === 'pm' ? 3600000 * 12 : 0
-                if (i < 4) {
-                    timeStrBefore = "12"
-                }
-                else {
-                    timeStrBefore = Math.floor(i / 4).toFixed().toString()
-                }
-
-                timeStrAfter = (function() {
-                    const decimals = fractionStr.substring(fractionStr.length - 2)
-                    return (decimals === '00') ? '00' :
-                        (decimals === '25') ? '15' :
-                        (decimals === '50') ? '30' :
-                        (decimals === '75') ? '45' :
-                        '00'
-                }())
-
-                timeStr = noAmPm ? timeStrBefore + ':' + timeStrAfter : timeStrBefore + ':' + timeStrAfter + ' ' + amPm
-                options.push({
-                    name: timeStr,
-                    value: (i * 900000) + halfDay,
-                })
-            }
-        }
-        loop('am')
-        loop('pm')
-        return options
-    }
-
-    public getTimesMap(): any {
-            const times = this.getTimes()
-            const map = {}
-            times.forEach(time => {
-                    map[time.value] = time.name
-            })
-            return map
     }
 }
