@@ -1,4 +1,3 @@
 export const environment = {
     staging: true,
-};
-  
+}
