@@ -1,7 +1,8 @@
 export const appConfig = {
+    brand_name: "Hyzer Shop",
     aws_bucket: "a2ee",
     blog_root: "blog",
-	client_url: "http://localhost:3000",
-	cloudfront_url: "https://d1eqpdomqeekcv.cloudfront.net",
-	paypal_env: 'sandbox',
+    client_url: "http://localhost:3000",
+    cloudfront_url: "https://d1eqpdomqeekcv.cloudfront.net",
+    paypal_env: 'sandbox',
 }
