@@ -1,5 +1,4 @@
 export default {
     MONGOOSE_USERNAME_FIELD: "email",
     PASSWORD_SALT_KEYLEN: 256,
-  };
-  
+}
