@@ -1,5 +1,6 @@
-export * from './constants';
-export * from './inversify';
+export * from './constants'
+export * from './inversify'
+export * from './currency'
 
-import { CONSTANTS } from './constants';
-export default CONSTANTS;
+import { CONSTANTS } from './constants'
+export default CONSTANTS
