@@ -2,6 +2,6 @@ export const TAGS = {
     AppController: 'AppController',
     UserController: 'UserController',
     ProductsController: 'ProductsController',
-};
+}
 
-export default TAGS;
+export default TAGS
