@@ -1,66 +1,68 @@
 import { Document, Types } from 'mongoose'
+import { Currency } from '../../constants/currency'
+import { IPrice } from './price'
 
 export interface IProduct extends Document {
 	/* Aesthetic */
-	name: string
-	slug: string
-	description?: string
-	featuredImages?: string[]
-	images?: string[]
-	largeImages?: string[]
-	thumbnails?: string[]
+    name: string
+    slug: string
+    description?: string
+    featuredImages?: string[]
+    images?: string[]
+    largeImages?: string[]
+    thumbnails?: string[]
 
 	/* Technical */
-	SKU: string
-	price?: number
-	priceRange?: number[]
-	salePrice?: number
-	salePriceRange?: number[]
-	onSale?: boolean
-	class: 'standalone' | 'parent' | 'variation'
-	isStandalone?: boolean
-	isParent?: boolean // Defines an abstract parent for a variable product
-	variationSKUs?: string[] // Array of product slugs
-	isVariation?: boolean // Defines a product variation with a parent product
-	parentSKU?: string // The slug of the parent product
+    SKU: string
+    price?: IPrice
+    priceRange?: IPrice[]
+    salePrice?: IPrice
+    salePriceRange?: IPrice[]
+    onSale?: boolean
+    class: 'standalone' | 'parent' | 'variation'
+    isStandalone?: boolean
+    isParent?: boolean // Defines an abstract parent for a variable product
+    variationSKUs?: string[] // Array of product slugs
+    isVariation?: boolean // Defines a product variation with a parent product
+    parentSKU?: string // The slug of the parent product
 
 	/* Attributes */
-	attributeValues?: Array<{
-		attributeId?: Types.ObjectId | string
-		attribute?: string
-		valueId?: Types.ObjectId | string
-		value?: any
-	}>
-	variableAttributes?: string[]
-	variableAttributeValues?: Array<{
-		attributeId?: Types.ObjectId | string
-		attribute?: string
-		valueId?: Types.ObjectId | string
-		value?: any
-	}>
-	
+    attributeValues?: Array<{
+        attributeId?: Types.ObjectId | string
+        attribute?: string
+        valueId?: Types.ObjectId | string
+        value?: any,
+    }>
+    variableAttributes?: string[]
+    variableAttributeValues?: Array<{
+        attributeId?: Types.ObjectId | string
+        attribute?: string
+        valueId?: Types.ObjectId | string
+        value?: any,
+    }>
+
 	/* Taxonomy */
-	taxonomyTerms: Array<Types.ObjectId | string>
-	taxonomyTermSlugs: string[]
+    taxonomyTerms: Array<Types.ObjectId | string>
+    taxonomyTermSlugs: string[]
 
 	/* Shipping */
-	units?: { // Only if not using global defaults
-		weight?: string
-		length?: string
-	}
-	dimensions?: {
-		length?: number
-		width?: number
-		height?: number
-	}
-	shippingWeight?: number
-	netWeight?: number
+    units?: { // Only if not using global defaults
+        weight?: string
+        length?: string,
+    }
+    dimensions?: {
+        length?: number
+        width?: number
+        height?: number,
+    }
+    shippingWeight?: number
+    netWeight?: number
 
 	/* Additional tax */
-	additionalTax?: number
+    additionalTax?: number
 
 	/* Sales */
-	stockQuantity?: number
-	totalSales?: number
-	isEnteredIntoStripe?: boolean
+    stockQuantity?: number
+    totalSales?: number
+    isEnteredIntoStripe?: boolean
 }
